+++ test/data/list-extra-async.py
@@ -1,13 +1,13 @@
 import logging
 import asyncclick as click
-from click_loglevel import LogLevel
+from click_loglevel.asyncclick import AsyncLogLevel
 
 logging.addLevelName(15, "VERBOSE")
 logging.addLevelName(25, "NOTICE")
 
 
 @click.command()
-@click.option("-l", "--log-level", type=LogLevel(extra=["VERBOSE", "NOTICE"]))
+@click.option("-l", "--log-level", type=AsyncLogLevel(extra=["VERBOSE", "NOTICE"]))
 async def main(log_level: int) -> None:
     click.echo(repr(log_level))