+++ test/test_with_asyncclick.py
@@ -6,19 +6,19 @@ import sys
 import asyncclick as click
 from asyncclick.testing import CliRunner
 import pytest
-from click_loglevel import LogLevel
+from click_loglevel.asyncclick import AsyncLogLevel
 
 DATA_DIR = Path(__file__).with_name("data")
 
 
 @click.command()
-@click.option("-l", "--log-level", type=LogLevel(), default=logging.INFO)
+@click.option("-l", "--log-level", type=AsyncLogLevel(), default=logging.INFO)
 async def lvlcmd(log_level: int) -> None:
     click.echo(repr(log_level))
 
 
 @click.command()
-@click.option("-l", "--log-level", type=LogLevel())
+@click.option("-l", "--log-level", type=AsyncLogLevel())
 async def lvlcmd_nodefault(log_level: int) -> None:
     click.echo(repr(log_level))
 
@@ -76,7 +76,6 @@ async def test_loglevel_no_default() -> None:
 
 
 @pytest.mark.anyio
-@pytest.mark.xfail(True, reason="asyncclick doesn't honor get_metavar")
 async def test_loglevel_help() -> None:
     r = await CliRunner().invoke(lvlcmd, ["--help"])
     assert r.exit_code == 0, r.output
@@ -95,7 +94,7 @@ async def test_loglevel_help() -> None:
 async def test_invalid_loglevel(value: str) -> None:
     r = await CliRunner().invoke(lvlcmd, ["--log-level", value])
     assert r.exit_code != 0, r.output
-    assert f"{value!r}: invalid log level" in str(r)
+    assert f"{value!r}: invalid log level" in r.output
 
 
 @pytest.mark.parametrize(
@@ -121,7 +120,6 @@ def test_loglevel_extra(loglevel: str | int, value: int, script: str) -> None:
     assert r.stdout == str(value) + "\n"
 
 
-@pytest.mark.xfail(True, reason="asyncclick doesn't honor get_metavar")
 @pytest.mark.parametrize("script", ["list-extra-async.py", "dict-extra-async.py"])
 def test_loglevel_extra_help(script: str) -> None:
     r = subprocess.run(
@@ -185,7 +183,6 @@ def test_loglevel_extra_nonupper(loglevel: str | int, value: int, script: str) -
     assert r.stdout == str(value) + "\n"
 
 
-@pytest.mark.xfail(True, reason="asyncclick doesn't honor get_metavar")
 @pytest.mark.parametrize(
     "script",
     [